Cleanup josh-core and remove old code
* split submodule and link handling out of lib.rs
* remove old leftover ACL code
