+++ josh-core/src/lib.rs
@@ -1,7 +1,5 @@
 #![warn(unused_extern_crates)]
 
-use anyhow::{Context, anyhow};
-
 #[macro_export]
 macro_rules! some_or {
     ($e:expr, $b:block) => {
@@ -24,7 +22,9 @@ pub mod changes;
 pub mod filter;
 pub mod history;
 pub mod housekeeping;
+pub mod link;
 pub mod shell;
+pub mod submodules;
 
 pub struct Change {
     pub author: String,
@@ -280,422 +280,4 @@ pub fn normalize_path(path: &std::path::Path) -> std::path::PathBuf {
         }
     }
     ret
-}
-
-type Users = std::collections::HashMap<String, User>;
-
-#[derive(Debug, serde::Deserialize)]
-struct User {
-    pub groups: Vec<String>,
-}
-
-type Groups = std::collections::HashMap<String, std::collections::HashMap<String, Group>>;
-#[derive(Debug, serde::Deserialize)]
-struct Group {
-    pub whitelist: String,
-    pub blacklist: String,
-}
-
-pub fn get_acl(
-    users: &str,
-    groups: &str,
-    user: &str,
-    repo: &str,
-) -> anyhow::Result<(filter::Filter, filter::Filter)> {
-    let users = std::fs::read_to_string(users).context("failed to read users file")?;
-    let users: Users = serde_yaml::from_str(&users).context("failed to parse users file")?;
-    let groups = std::fs::read_to_string(groups).context("failed to read groups file")?;
-    let groups: Groups = serde_yaml::from_str(&groups).context("failed to parse groups file")?;
-
-    users
-        .get(user)
-        .and_then(|u| {
-            let mut whitelist = filter::Filter::new().empty();
-            let mut blacklist = filter::Filter::new().empty();
-            for g in &u.groups {
-                let lists = groups.get(repo).and_then(|repo| {
-                    repo.get(g.as_str()).map(|group| {
-                        let w = filter::parse(&group.whitelist);
-                        let b = filter::parse(&group.blacklist);
-                        (w, b)
-                    })
-                })?;
-                if let Err(e) = lists.0 {
-                    return Some(Err(anyhow!("Error parsing whitelist: {}", e)));
-                }
-                if let Err(e) = lists.1 {
-                    return Some(Err(anyhow!("Error parsing blacklist: {}", e)));
-                }
-                if let Ok(w) = lists.0 {
-                    whitelist = filter::compose(&[whitelist, w]);
-                }
-                if let Ok(b) = lists.1 {
-                    blacklist = filter::compose(&[blacklist, b]);
-                }
-            }
-            println!("w: {:?}, b: {:?}", whitelist, blacklist);
-            Some(Ok((whitelist, blacklist)))
-        })
-        .unwrap_or_else(|| Ok((filter::Filter::new().empty(), filter::Filter::new())))
-}
-
-#[cfg(feature = "incubating")]
-pub struct ParsedSubmoduleEntry {
-    pub path: std::path::PathBuf,
-    pub url: String,
-    pub branch: String,
-}
-
-#[cfg(feature = "incubating")]
-pub fn parse_gitmodules(gitmodules_content: &str) -> anyhow::Result<Vec<ParsedSubmoduleEntry>> {
-    use gix_submodule::File;
-
-    let submodules = File::from_bytes(gitmodules_content.as_bytes(), None, &Default::default())
-        .context("Failed to parse .gitmodules")?;
-
-    let mut entries: Vec<ParsedSubmoduleEntry> = Vec::new();
-
-    for name in submodules.names() {
-        // path is required to consider an entry
-        if let Ok(path) = submodules.path(name) {
-            let path = std::path::PathBuf::from(path.to_string());
-
-            let url = submodules
-                .url(name)
-                .ok()
-                .map(|u| u.to_string())
-                .unwrap_or_default();
-
-            // Default branch to "HEAD" if not configured
-            let branch = submodules
-                .branch(name)
-                .ok()
-                .and_then(|opt| {
-                    opt.map(|b| match b {
-                        gix_submodule::config::Branch::CurrentInSuperproject => ".".to_string(),
-                        gix_submodule::config::Branch::Name(n) => n.to_string(),
-                    })
-                })
-                .unwrap_or_else(|| "HEAD".to_string());
-
-            entries.push(ParsedSubmoduleEntry { path, url, branch });
-        }
-    }
-
-    Ok(entries)
-}
-
-#[cfg(feature = "incubating")]
-pub fn update_gitmodules(
-    gitmodules_content: &str,
-    entry: &ParsedSubmoduleEntry,
-) -> anyhow::Result<String> {
-    use gix_config::File as ConfigFile;
-    use gix_submodule::File as SubmoduleFile;
-
-    // Parse the existing gitmodules content using gix_submodule
-    let submodule_file = SubmoduleFile::from_bytes(
-        gitmodules_content.as_bytes(),
-        None,
-        &ConfigFile::new(gix_config::file::Metadata::default()),
-    )
-    .context("Failed to parse .gitmodules")?;
-
-    // Get the underlying config file to modify it
-    let mut config = submodule_file.config().clone();
-
-    // Find the existing submodule by matching the path
-    let mut existing_submodule_name = None;
-    for name in submodule_file.names() {
-        if let Ok(path) = submodule_file.path(name) {
-            if path.to_string() == entry.path.to_string_lossy() {
-                existing_submodule_name = Some(name.to_string());
-                break;
-            }
-        }
-    }
-
-    let submodule_name = if let Some(name) = existing_submodule_name {
-        // Use the existing submodule name
-        name
-    } else {
-        // Create a new submodule name from path (fallback)
-        entry.path.to_string_lossy().replace('/', "_")
-    };
-
-    // Create or update the submodule section
-    let mut section = config
-        .section_mut_or_create_new("submodule", Some(submodule_name.as_str().into()))
-        .context("Failed to create submodule section")?;
-
-    // Remove existing keys if they exist to avoid duplicates
-    use gix_config::parse::section::ValueName;
-
-    // Remove all existing values for these keys
-    while section.remove("path").is_some() {}
-    while section.remove("url").is_some() {}
-    while section.remove("branch").is_some() {}
-
-    // Set the submodule properties using push method
-    let path_key: ValueName = "path".try_into()?;
-    let url_key: ValueName = "url".try_into()?;
-    let branch_key: ValueName = "branch".try_into()?;
-
-    section.push(path_key, Some(entry.path.to_string_lossy().as_ref().into()));
-    section.push(url_key, Some(entry.url.as_str().into()));
-    if entry.branch != "HEAD" {
-        section.push(branch_key, Some(entry.branch.as_str().into()));
-    }
-
-    // Write the updated config back to string
-    let mut output = Vec::new();
-    config
-        .write_to(&mut output)
-        .context("Failed to write gitmodules")?;
-
-    String::from_utf8(output).context("Invalid UTF-8 in gitmodules")
-}
-
-#[cfg(feature = "incubating")]
-pub fn find_link_files(
-    repo: &git2::Repository,
-    tree: &git2::Tree,
-) -> anyhow::Result<Vec<(std::path::PathBuf, filter::Filter)>> {
-    let mut link_files = Vec::new();
-
-    tree.walk(git2::TreeWalkMode::PreOrder, |root, entry| {
-        if let Some(name) = entry.name() {
-            if name == ".link.josh" {
-                // Found a link file
-                let link_blob = match repo.find_blob(entry.id()) {
-                    Ok(blob) => blob,
-                    Err(e) => {
-                        eprintln!("Failed to find blob: {}", e);
-                        return git2::TreeWalkResult::Skip;
-                    }
-                };
-
-                let link_content = match std::str::from_utf8(link_blob.content()) {
-                    Ok(content) => content,
-                    Err(e) => {
-                        eprintln!("Failed to parse link file content: {}", e);
-                        return git2::TreeWalkResult::Skip;
-                    }
-                };
-
-                let filter = match filter::parse(link_content.trim()) {
-                    Ok(f) => f,
-                    Err(e) => {
-                        eprintln!("Failed to parse .link.josh filter: {}", e);
-                        return git2::TreeWalkResult::Skip;
-                    }
-                };
-
-                let root = root.trim_matches('/');
-                // Use root as the directory path where the .link.josh file is located
-                let path = std::path::PathBuf::from(root);
-
-                link_files.push((path, filter));
-            }
-        }
-
-        git2::TreeWalkResult::Ok
-    })
-    .context("Failed to walk tree")?;
-
-    Ok(link_files)
-}
-
-#[cfg(feature = "incubating")]
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use std::path::PathBuf;
-
-    #[test]
-    fn test_parse_gitmodules_basic() {
-        let content = r#"[submodule "libs/foo"]
-	path = libs/foo
-	url = https://github.com/example/foo.git
-	branch = main
-
-[submodule "libs/bar"]
-	path = libs/bar
-	url = https://github.com/example/bar.git"#;
-
-        let result = parse_gitmodules(content).unwrap();
-        assert_eq!(result.len(), 2);
-
-        assert_eq!(result[0].path, PathBuf::from("libs/foo"));
-        assert_eq!(result[0].url, "https://github.com/example/foo.git");
-        assert_eq!(result[0].branch, "main");
-
-        assert_eq!(result[1].path, PathBuf::from("libs/bar"));
-        assert_eq!(result[1].url, "https://github.com/example/bar.git");
-        assert_eq!(result[1].branch, "HEAD"); // default
-    }
-
-    #[test]
-    fn test_parse_gitmodules_empty() {
-        let content = "";
-        let result = parse_gitmodules(content).unwrap();
-        assert_eq!(result.len(), 0);
-    }
-
-    #[test]
-    fn test_parse_gitmodules_invalid() {
-        let content = "invalid gitmodules content";
-        let result = parse_gitmodules(content);
-        assert!(result.is_err());
-    }
-
-    #[test]
-    fn test_update_gitmodules_add_new() {
-        let content = "";
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("libs/foo"),
-            url: "https://github.com/example/foo.git".to_string(),
-            branch: "main".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry).unwrap();
-
-        let expected = r#"[submodule "libs_foo"]
-	path = libs/foo
-	url = https://github.com/example/foo.git
-	branch = main
-"#;
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_update_gitmodules_add_new_with_default_branch() {
-        let content = "";
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("libs/bar"),
-            url: "https://github.com/example/bar.git".to_string(),
-            branch: "HEAD".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry).unwrap();
-
-        let expected = r#"[submodule "libs_bar"]
-	path = libs/bar
-	url = https://github.com/example/bar.git
-"#;
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_update_gitmodules_update_existing() {
-        let content = r#"[submodule "existing_foo"]
-	path = libs/foo
-	url = https://github.com/example/old-foo.git
-	branch = old-branch"#;
-
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("libs/foo"),
-            url: "https://github.com/example/new-foo.git".to_string(),
-            branch: "new-branch".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry).unwrap();
-
-        // Should update existing values, not append duplicates
-        let expected = r#"[submodule "existing_foo"]
-	path = libs/foo
-	url = https://github.com/example/new-foo.git
-	branch = new-branch
-"#;
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_update_gitmodules_update_existing_with_default_branch() {
-        let content = r#"[submodule "existing_bar"]
-	path = libs/bar
-	url = https://github.com/example/old-bar.git
-	branch = old-branch"#;
-
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("libs/bar"),
-            url: "https://github.com/example/new-bar.git".to_string(),
-            branch: "HEAD".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry).unwrap();
-
-        // Should update existing values and remove branch when it's HEAD
-        let expected = r#"[submodule "existing_bar"]
-	path = libs/bar
-	url = https://github.com/example/new-bar.git
-"#;
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_update_gitmodules_multiple_submodules() {
-        let content = r#"[submodule "existing_foo"]
-	path = libs/foo
-	url = https://github.com/example/foo.git
-
-[submodule "existing_bar"]
-	path = libs/bar
-	url = https://github.com/example/bar.git"#;
-
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("libs/baz"),
-            url: "https://github.com/example/baz.git".to_string(),
-            branch: "develop".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry).unwrap();
-
-        // The gix-config API appends new sections at the end
-        let expected = r#"[submodule "existing_foo"]
-	path = libs/foo
-	url = https://github.com/example/foo.git
-
-[submodule "existing_bar"]
-	path = libs/bar
-	url = https://github.com/example/bar.git
-[submodule "libs_baz"]
-	path = libs/baz
-	url = https://github.com/example/baz.git
-	branch = develop
-"#;
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_update_gitmodules_path_with_slashes() {
-        let content = "";
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("deep/nested/path/submodule"),
-            url: "https://github.com/example/deep-submodule.git".to_string(),
-            branch: "main".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry).unwrap();
-
-        let expected = r#"[submodule "deep_nested_path_submodule"]
-	path = deep/nested/path/submodule
-	url = https://github.com/example/deep-submodule.git
-	branch = main
-"#;
-        assert_eq!(result, expected);
-    }
-
-    #[test]
-    fn test_update_gitmodules_invalid_content() {
-        let content = "invalid gitmodules content";
-        let entry = ParsedSubmoduleEntry {
-            path: PathBuf::from("libs/foo"),
-            url: "https://github.com/example/foo.git".to_string(),
-            branch: "main".to_string(),
-        };
-
-        let result = update_gitmodules(content, &entry);
-        assert!(result.is_err());
-    }
 }
